refactor: use absolute imports in __init__.py and rename format_adapter to message_models
